refactor: change `duration` column type to `INT`

+++ admin-writer/src/main/kotlin/com/epam/drill/admin/writer/rawdata/entity/TestMetadata.kt
@@ -28,7 +28,7 @@ class TestLaunch (
     val testDefinitionId: String,
     val testSessionId: String,
     val result: String?,
-    val duration: Long? = null,
+    val duration: Int? = null,
 )
 
 class TestDefinition(
